+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - Replace `--no_overwrite` with `--exists` in `zarrs_filter` and `zarrs_ome`
+   - Both support `erase` and `exit` options
+   - `zarrs_ome` also supports an `overwrite` option
+
 ## [0.4.0] - 2024-04-20
 
 ### Added